32bit status, and full maxid on store

+++ buoy/ads1282.h
@@ -233,7 +233,7 @@ namespace Buoy {
 
       volatile uint64_t references  [BATCHES];
       volatile uint64_t microdeltas [BATCHES];
-      volatile uint8_t  reference_status [BATCHES];
+      volatile uint32_t  reference_status [BATCHES];
 
       volatile uint32_t totalsamples;
